requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
psycopg[binary]
pytest
reportlab
gunicorn
